[1.1.0-beta] X64 Hooking, MemoryBuffer Upgrade, Heaps of Changes
Remove Redundant Code
Refactor Method Names
Remove Old Commented Out Code
Fix Minor Incorrect Calling Convention Wrapper Assembly Behaviour
Refactor Position of Classes
Utilities to Interact and Manage Memory Pages
X64 Sample Mods
Assembler will no longer disconnect during debugging.

+++ libReloaded/Process/Memory/Allocate.cs
@@ -43,7 +43,7 @@ namespace Reloaded.Process.Memory
             (
                 process.ProcessHandle, 
                 IntPtr.Zero, 
-                (IntPtr)length, 
+                (uint)length, 
                 AllocationTypes.Commit | AllocationTypes.Reserve, 
                 MemoryProtections.ExecuteReadWrite
             );
